CLN : literalinclude enum editor in factories basic
	new file:   docs/source/traitsui_user_manual/examples/enum_editor.py
	modified:   docs/source/traitsui_user_manual/factories_basic.rst
	deleted:    examples/tutorials/doc_examples/examples/enum_editor.py
